move keymap setup into separate file
